Update README.md
I stumbled over the situation where I was in an existing index.ts file (with the file selected in the tree view) and the command didn't worked. After I looked in the code it was clear: `getActiveFilePath` returned the path to the index.ts file and this is not a directory. Solution was to select the directory in the tree view.